Ready for IOS

+++ lib/c_protocols/note_protocols/fire/note_fire_ops.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:bldrs/a_models/e_notes/a_note_model.dart';
 import 'package:fire/super_fire.dart';
 import 'package:bldrs/e_back_end/b_fire/foundation/fire_paths.dart';
-import 'package:mapper/mapper.dart';
+import 'package:basics/helpers/classes/maps/mapper.dart';
 import 'package:flutter/material.dart';
 
 class NoteFireOps {
@@ -16,19 +16,19 @@ class NoteFireOps {
 
   // --------------------
   /// TESTED : WORKS PERFECT
-  static Future<NoteModel> createNote({
-    @required NoteModel noteModel,
-    ValueChanged<NoteModel> onFinished,
+  static Future<NoteModel?> createNote({
+    required NoteModel? noteModel,
+    ValueChanged<NoteModel?>? onFinished,
   }) async {
-    NoteModel _output;
+    NoteModel? _output;
 
     if (noteModel != null){
 
-      if (noteModel.sendNote == true){
+      if (noteModel.sendNote != null && noteModel.sendNote! == true){
 
-        final String docID = await Fire.createDoc(
-          coll: FireColl.getPartyCollName(noteModel.parties.receiverType),
-          doc: noteModel.parties.receiverID,
+        final String? docID = await Fire.createDoc(
+          coll: FireColl.getPartyCollName(noteModel.parties?.receiverType),
+          doc: noteModel.parties?.receiverID,
           subColl: FireSubColl.noteReceiver_receiver_notes,
           input: noteModel.toMap(toJSON: false),
         );
@@ -56,8 +56,8 @@ class NoteFireOps {
   // --------------------
   ///
   static Future<List<NoteModel>> createNotes({
-    @required NoteModel noteModel,
-    @required List<String> receiversIDs,
+    required NoteModel? noteModel,
+    required List<String>? receiversIDs,
   }) async {
 
     final List<NoteModel> _output = <NoteModel>[];
@@ -67,17 +67,17 @@ class NoteFireOps {
 
       await Future.wait(<Future>[
 
-        ...List.generate(receiversIDs.length, (index){
+        ...List.generate(receiversIDs!.length, (index){
 
           final NoteModel _note = noteModel.copyWith(
-            parties: noteModel.parties.copyWith(
+            parties: noteModel.parties?.copyWith(
               receiverID: receiversIDs[index],
             ),
           );
 
           return createNote(
             noteModel: _note,
-            onFinished: (NoteModel uploaded){
+            onFinished: (NoteModel? uploaded){
               if (uploaded != null){
                 _output.add(uploaded);
               }
@@ -91,7 +91,7 @@ class NoteFireOps {
       _success = true;
     }
 
-    return _success == true ? _output : null;
+    return _success == true ? _output : [];
   }
   // -----------------------------------------------------------------------------
 
@@ -99,15 +99,15 @@ class NoteFireOps {
 
   // --------------------
   /// TESTED : WORKS PERFECT
-  static Future<NoteModel> readNote({
-    @required String noteID,
-    @required String userID,
+  static Future<NoteModel?> readNote({
+    required String? noteID,
+    required String? userID,
   }) async {
-    NoteModel _output;
+    NoteModel? _output;
 
     if (noteID != null && userID != null){
 
-      final Map<String, dynamic> map = await Fire.readDoc(
+      final Map<String, dynamic>? map = await Fire.readDoc(
         coll: FireColl.users,
         doc: userID,
         subColl: FireSubColl.noteReceiver_receiver_notes,
@@ -131,14 +131,14 @@ class NoteFireOps {
   // --------------------
   /// TESTED : WORKS PERFECT
   static Future<void> updateNote({
-    @required NoteModel note,
+    required NoteModel? note,
   }) async {
 
-    if (note != null){
+    if (note != null && note.parties?.receiverID != null){
 
       await Fire.updateDoc(
-        coll: FireColl.getPartyCollName(note.parties.receiverType),
-        doc: note.parties.receiverID,
+        coll: FireColl.getPartyCollName(note.parties?.receiverType),
+        doc: note.parties!.receiverID!,
         subColl: FireSubColl.noteReceiver_receiver_notes,
         subDoc: note.id,
         input: note.toMap(toJSON: false),
@@ -149,7 +149,7 @@ class NoteFireOps {
   // --------------------
   /// TESTED : WORKS PERFECT
   static Future<void> markNoteAsSeen({
-    @required NoteModel noteModel,
+    required NoteModel noteModel,
   }) async {
 
     if (noteModel.seen != true){
@@ -170,13 +170,13 @@ class NoteFireOps {
   // --------------------
   ///
   static Future<void> markNotesAsSeen({
-    @required List<NoteModel> notes,
+    required List<NoteModel>? notes,
   }) async {
 
     if (Mapper.checkCanLoopList(notes) == true){
 
       /// MARK ON FIREBASE
-      for (final NoteModel note in notes){
+      for (final NoteModel note in notes!){
         unawaited(markNoteAsSeen(
           noteModel: note,
         ));
@@ -193,14 +193,14 @@ class NoteFireOps {
   // --------------------
   /// TESTED : WORKS PERFECT
   static Future<void> deleteNote({
-    @required NoteModel note,
+    required NoteModel? note,
   }) async {
 
-    if (note != null){
+    if (note != null && note.parties?.receiverID != null){
 
       await Fire.deleteDoc(
-        coll: FireColl.getPartyCollName(note.parties.receiverType),
-        doc: note.parties.receiverID,
+        coll: FireColl.getPartyCollName(note.parties?.receiverType),
+        doc: note.parties!.receiverID!,
         subColl: FireSubColl.noteReceiver_receiver_notes,
         subDoc: note.id,
       );
@@ -211,14 +211,14 @@ class NoteFireOps {
   // --------------------
   ///
   static Future<void> deleteNotes({
-    @required List<NoteModel> notes,
+    required List<NoteModel>? notes,
   }) async {
 
     if (Mapper.checkCanLoopList(notes) == true){
 
       await Future.wait(<Future>[
 
-        ...List.generate(notes.length, (index){
+        ...List.generate(notes!.length, (index){
 
           return deleteNote(
             note: notes[index],
