+++ lib/c_protocols/pic_protocols/storage/pic_storage_ops.dart
@@ -1,12 +1,14 @@
 import 'dart:typed_data';
-
+import 'package:basics/helpers/classes/checks/object_check.dart';
+import 'package:basics/helpers/classes/checks/tracers.dart';
+import 'package:basics/helpers/classes/files/file_size_unit.dart';
+import 'package:basics/helpers/classes/files/floaters.dart';
+import 'package:basics/helpers/classes/strings/text_check.dart';
 import 'package:bldrs/a_models/i_pic/pic_model.dart';
 import 'package:fire/super_fire.dart';
-import 'package:filers/filers.dart';
-import 'package:flutter/material.dart';
-import 'package:mapper/mapper.dart';
-import 'package:mediators/mediators.dart';
-import 'package:stringer/stringer.dart';
+import 'package:basics/helpers/classes/files/filers.dart';
+import 'package:basics/helpers/classes/maps/mapper.dart';
+import 'package:mediators/models/dimension_model.dart';
 
 class PicStorageOps {
   // -----------------------------------------------------------------------------
@@ -19,14 +21,14 @@ class PicStorageOps {
 
   // --------------------
   /// TESTED: WORKS PERFECT
-  static Future<PicModel> createPic(PicModel picModel) async {
+  static Future<PicModel?> createPic(PicModel? picModel) async {
 
     PicModel.assertIsUploadable(picModel);
 
-    final String _url = await Storage.uploadBytesAndGetURL(
-      bytes: picModel.bytes,
-      path: picModel.path,
-      storageMetaModel: picModel.meta,
+    final String? _url = await Storage.uploadBytesAndGetURL(
+      bytes: picModel?.bytes,
+      path: picModel?.path,
+      storageMetaModel: picModel?.meta,
     );
 
     if (_url == null){
@@ -44,16 +46,16 @@ class PicStorageOps {
 
   // --------------------
   /// TESTED : WORKS PERFECT
-  static Future<PicModel> readPic({
-    @required String path,
+  static Future<PicModel?> readPic({
+    required String? path,
   }) async {
-    PicModel _picModel;
+    PicModel? _picModel;
 
     if (TextCheck.isEmpty(path) == false){
 
       final bool _pathIsURL = ObjectCheck.isAbsoluteURL(path);
-      Uint8List _bytes;
-      StorageMetaModel _meta;
+      Uint8List? _bytes;
+      StorageMetaModel? _meta;
 
       /// GET BYTES
       if (_pathIsURL == true){
@@ -72,13 +74,13 @@ class PicStorageOps {
         );
       }
       else if (_pathIsURL == true){
-      final Dimensions _dims = await Dimensions.superDimensions(_bytes);
-      final double _mega = Filers.calculateSize(_bytes.length, FileSizeUnit.megaByte);
+      final Dimensions? _dims = await Dimensions.superDimensions(_bytes);
+      final double? _mega = Filers.calculateSize(_bytes?.length, FileSizeUnit.megaByte);
         _meta = StorageMetaModel(
           ownersIDs: const ['non'],
           name: path,
-          height: _dims.height,
-          width: _dims.width,
+          height: _dims?.height,
+          width: _dims?.width,
           sizeMB: _mega,
         );
       }
@@ -99,11 +101,11 @@ class PicStorageOps {
 
   // --------------------
   /// TASK : TEST ME
-  static Future<PicModel> updatePic({
-    @required PicModel picModel,
+  static Future<PicModel?> updatePic({
+    required PicModel? picModel,
   }) async {
 
-    final PicModel _uploaded = await createPic(picModel);
+    final PicModel? _uploaded = await createPic(picModel);
 
     return _uploaded;
   }
@@ -113,25 +115,27 @@ class PicStorageOps {
 
   // --------------------
   /// TASK : TEST ME
-  static Future<void> deletePic(String path) async {
+  static Future<void> deletePic(String? path) async {
     blog('deletePic : START');
 
-    await Storage.deleteDoc(
-      path: path,
-      currentUserID: Authing.getUserID(),
-    );
+    if (path != null && Authing.getUserID() != null){
+      await Storage.deleteDoc(
+        path: path,
+        currentUserID: Authing.getUserID()!,
+      );
+    }
 
     blog('deletePic : END');
   }
   // --------------------
   /// TASK : TEST ME
-  static Future<void> deletePics(List<String> paths) async {
+  static Future<void> deletePics(List<String>? paths) async {
 
-    if (Mapper.checkCanLoopList(paths) == true){
+    if (Authing.getUserID() != null && Mapper.checkCanLoopList(paths) == true){
 
       await Storage.deleteDocs(
-        paths: paths,
-        currentUserID: Authing.getUserID(),
+        paths: paths!,
+        currentUserID: Authing.getUserID()!,
       );
 
     }
